:bug: fix rvm failing on update
trying to run:

  `rvm get stable`

in the environment of the update script was failing because `rvm` didn't
technically exist (this may indicate a larger problem in that the update
tool `udot` may be to blame here.  It is specially located to allow for
execution with sudo and this may be screwing up the scripts ability to
find the rvm executable). For now though, I've added home path relative
path for the rvm executable.